tests/wpt/tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.get.rounding.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.imageData.get.rounding</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.get.rounding</h1>
<p class="desc">Test the handling of non-integer source coordinates in getImageData().</p>


<script>
var t = async_test("Test the handling of non-integer source coordinates in getImageData().");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  function testDimensions(sx, sy, sw, sh, width, height)
  {
      imageData = ctx.getImageData(sx, sy, sw, sh);
      _assert(imageData.width == width, "imageData.width == width");
      _assert(imageData.height == height, "imageData.height == height");
  }

  testDimensions(0, 0, 20, 10, 20, 10);

  testDimensions(.1, .2, 20, 10, 20, 10);
  testDimensions(.9, .8, 20, 10, 20, 10);

  testDimensions(0, 0, 20.9, 10.9, 20, 10);
  testDimensions(0, 0, 20.1, 10.1, 20, 10);

  testDimensions(-1, -1, 20, 10, 20, 10);

  testDimensions(-1.1, 0, 20, 10, 20, 10);
  testDimensions(-1.9,  0, 20, 10, 20, 10);
  t.done();

});
</script>
